move glibc-specific stuff into the glibc file
