+++ pwndbg/commands/__init__.py
@@ -27,12 +27,7 @@ import pwndbg.color
 import pwndbg.dbg_mod
 import pwndbg.dintegration
 import pwndbg.exception
-import pwndbg.libc
-from pwndbg.aglib.heap.ptmalloc import DebugSymsHeap
-from pwndbg.aglib.heap.ptmalloc import GlibcMemoryAllocator
-from pwndbg.aglib.heap.ptmalloc import HeuristicHeap
 from pwndbg.color import message
-from pwndbg.lib import SymbolNotRecoveredError
 from pwndbg.lib import TypeNotRecoveredError
 
 log = logging.getLogger(__name__)
@@ -864,139 +859,6 @@ def OnlyWhenRunning(
     func_when_no_kwargs._pwndbg_only_when_running_allow_core = allow_core  # type: ignore[attr-defined]
     return func_when_no_kwargs
 
-
-def OnlyWithTcache(function: Callable[P, T]) -> Callable[P, T | None]:
-    @functools.wraps(function)
-    def _OnlyWithTcache(*a: P.args, **kw: P.kwargs) -> T | None:
-        assert isinstance(pwndbg.aglib.heap.current, GlibcMemoryAllocator)
-        if pwndbg.aglib.heap.current.has_tcache():
-            return function(*a, **kw)
-        log.error(
-            f"{func_name(function)}: This version of GLIBC was not compiled with tcache support."
-        )
-        return None
-
-    return _OnlyWithTcache
-
-
-def OnlyWhenHeapIsInitialized(function: Callable[P, T]) -> Callable[P, T | None]:
-    @functools.wraps(function)
-    def _OnlyWhenHeapIsInitialized(*a: P.args, **kw: P.kwargs) -> T | None:
-        if pwndbg.aglib.heap.current is not None and pwndbg.aglib.heap.current.is_initialized():
-            return function(*a, **kw)
-        log.error(f"{func_name(function)}: Heap is not initialized yet.")
-        return None
-
-    return _OnlyWhenHeapIsInitialized
-
-
-def _try2run_heap_command(function: Callable[P, T], *a: P.args, **kw: P.kwargs) -> T | None:
-    e = log.error
-    w = log.warning
-    # Note: We will still raise the error for developers when exception-* is set to "on"
-    try:
-        return function(*a, **kw)
-    except SymbolNotRecoveredError as err:
-        e(f"{func_name(function)}: Fail to resolve the symbol: `{err.name}`")
-        if "thread_arena" == err.name:
-            w(
-                "You are probably debugging a multi-threaded target without debug symbols, so we failed to determine which arena is used by the current thread.\n"
-                "To resolve this issue, you can use the `arenas` command to list all arenas, and use `set thread-arena <addr>` to set the current thread's arena address you think is correct.\n"
-            )
-        else:
-            w(
-                f"You can try to determine the libc symbols addresses manually and set them appropriately. For this, see the `heap-config` command output and set the config for `{err.name}`."
-            )
-        if pwndbg.config.exception_verbose or pwndbg.config.exception_debugger:
-            raise err
-
-        pwndbg.exception.inform_verbose_and_debug()
-    except Exception as err:
-        e(f"{func_name(function)}: An unknown error occurred when running this command.")
-        if isinstance(pwndbg.aglib.heap.current, HeuristicHeap):
-            w(
-                "Maybe you can try to determine the libc symbols addresses manually, set them appropriately and re-run this command. For this, see the `heap-config` command output and set the `main_arena`, `mp_`, `global_max_fast`, `tcache` and `thread_arena` addresses."
-            )
-        else:
-            w("You can try `set resolve-heap-via-heuristic force` and re-run this command.\n")
-        if pwndbg.config.exception_verbose or pwndbg.config.exception_debugger:
-            raise err
-
-        pwndbg.exception.inform_verbose_and_debug()
-    return None
-
-
-def OnlyWithResolvedHeapSyms(function: Callable[P, T]) -> Callable[P, T | None]:
-    @functools.wraps(function)
-    def _OnlyWithResolvedHeapSyms(*a: P.args, **kw: P.kwargs) -> T | None:
-        e = log.error
-        w = log.warning
-
-        # Operating under the assumption that the pwndbg/libc/ code can figure out
-        # that we are using glibc with at least as good accuracy as the ptmalloc code.
-        if pwndbg.libc.which() != pwndbg.libc.LibcType.GLIBC:
-            e(f"The currently active libc isn't glibc. It's {pwndbg.libc.which().value}.")
-            return None
-
-        if (
-            isinstance(pwndbg.aglib.heap.current, HeuristicHeap)
-            and pwndbg.config.resolve_heap_via_heuristic == "auto"
-            and DebugSymsHeap().can_be_resolved()
-        ):
-            # In auto mode, we will try to use the debug symbols if possible
-            pwndbg.aglib.heap.current = DebugSymsHeap()
-
-        if (
-            pwndbg.aglib.heap.current is not None
-            and isinstance(pwndbg.aglib.heap.current, GlibcMemoryAllocator)
-            and pwndbg.aglib.heap.current.can_be_resolved()
-        ):
-            return _try2run_heap_command(function, *a, **kw)
-
-        static = not pwndbg.dbg.selected_inferior().is_dynamically_linked()
-        if (
-            isinstance(pwndbg.aglib.heap.current, DebugSymsHeap)
-            and pwndbg.config.resolve_heap_via_heuristic == "auto"
-        ):
-            # In auto mode, if the debug symbols are not enough, we will try to use the heuristic if possible
-            heuristic_heap = HeuristicHeap()
-            if heuristic_heap.can_be_resolved():
-                pwndbg.aglib.heap.current = heuristic_heap
-                w(
-                    "pwndbg will try to resolve the heap symbols via heuristic now since we cannot resolve the heap via the debug symbols.\n"
-                    "This might not work in all cases. Use `help set resolve-heap-via-heuristic` for more details.\n"
-                )
-                return _try2run_heap_command(function, *a, **kw)
-            if static:
-                e(
-                    "Can't find GLIBC version required for this command to work since this is a statically linked binary"
-                )
-                w(
-                    "Please set the GLIBC version you think the target binary was compiled (using `set glibc <version>` command; e.g. 2.32) and re-run this command."
-                )
-            else:
-                e(
-                    "Can't find GLIBC version required for this command to work, maybe is because GLIBC is not loaded yet."
-                )
-                w(
-                    "If you believe the GLIBC is loaded or this is a statically linked binary. "
-                    "Please set the GLIBC version you think the target binary was compiled (using `set glibc <version>` command; e.g. 2.32) and re-run this command"
-                )
-        elif (
-            isinstance(pwndbg.aglib.heap.current, DebugSymsHeap)
-            and pwndbg.config.resolve_heap_via_heuristic == "force"
-        ):
-            e(
-                "You are forcing to resolve the heap symbols via heuristic, but we cannot resolve the heap via the debug symbols."
-            )
-            w("Use `set resolve-heap-via-heuristic auto` and re-run this command.")
-        else:
-            # Note: Should not see this error, but just in case
-            e("An unknown error occurred when resolved the heap.")
-            pwndbg.exception.inform_report_issue("An unknown error occurred when resolved the heap")
-        return None
-
-    return _OnlyWithResolvedHeapSyms
 
 
 def sloppy_gdb_parse(s: str) -> int | str:
